3.20180917.0821.43f8f4a861 (11 features, 43 changes)
Update to Rhino 1.7.9
Update dependencies

Update standard plugins to d05f827

Change database index to GIN for much better performance on large
applications. To update existing applications, run the migration script:

   db/migrations/20180826-use-gin-index.rb

User management UI allows representative objects on groups as well as
users
Object editor: /do/edit URL can have success_redirect/then parameter to
redirect after successful edit, just like the new object editor
template.
LDAP remote authentication server can be configured with a list of
attributes that should be queried from the server. Defaults to a
reasonable list covering a range of LDAP servers.
Add restrictToApplicationId to plugin.json files, which sets the plugin
to only be visible to that application so that multiple versions of the
same plugin can be run at once in deployment. Change reloading logic to
be aware of these plugins.
JS WorkUnit isActionableBy() allows groups to be passed it.
WorkUnitQuery ref() method throws an exception when null passed as
argument, because previous behaviour just caused annoying bugs. If you
don't want to filter by ref, don't call ref().
Add std:empty and std:clear layouts plus HSVT render functions
Object editor plugin adaptor: Add getContainerChoosesQualifier() to
containers so plugins can tell whether the attribute has qualifier UI.
Add loadVersion() and loadVersionAtTime() to JS Ref interface
Web publisher support updated to match plugin.
Render larger file thumbnail images, but display them at the same size
in the application. This makes them look better on HiDPI screens, and
allows bigger thumbnails to be displayed in web publisher publications.
std:file* template functions have additional
authenticationSignatureValidForSeconds option.
Add createShowType property to info returned from SCHEMA.getTypeInfo()
Add request argument to hPreFileDownload
Add platform component to read GeoLite2 database, exposed through
haplo_info_geoip plugin
Add hObjectRender and hObjectRenderPublisher hooks so plugins can hide
attributes without having to do things in hPreObjectDisplay which
conflict horribly with computed attributes.
More control over computed attributes, intended for optimisations and
plugins to control updates to objects.
JS HTTP client spills long responses to disk files for efficiency.
JS HTTP client's bodyParameter() and queryParameter() allow the value
to be null or undefined, in which case the value is treated as the
empty string.
Service users have 'not anonymous' policy so plugins which use them for
authentication don't need to allow anon requests.
Add a progress log message when doing a full text reindex.
Force the OAuth client to use TLS 1.2, to minimise issues with weird
SSL configs on servers.
Move location of components in source tree

oForms updated to 2d4f1d8727
 - preview button to files in documents
 - add externalData(), for include logic, render-template elements etc

Fixes
 - HSTORE parser fixed to cope with empty sets
 - previews of multi-page PDFs where the request is signed
 - Schema requirements: feature std:navigation entries allow _ in
names, and an error is output when a bad name is used.
 - System management UI allows deletion of all navigation entries.
 - Styling of lists
 - HTTP client: Use HSVT to encode URL arguments for correctness and to
meet expectations of other systems.
 - Remove redundant code
 - Work around differing date precision in object history retrieval APIs

--
fossil: 43f8f4a861
approved-by: ben

+++ test/integration/javascript_plugin_loading_test.rb
@@ -13,6 +13,7 @@ class JavaScriptPluginLoadingTest < IntegrationTest
   NEW_PLUGIN_SOURCE = "#{File.dirname(__FILE__)}/javascript/javascript_plugin_loading/new_3p_plugin"
   NEW_PLUGIN_DEST_DIR = "#{PLUGINS_LOCAL_DIRECTORY}/test"
   NEW_PLUGIN_DEST = "#{NEW_PLUGIN_DEST_DIR}/new_3p_plugin"
+  NEW_PLUGIN_DEST_APP_RESTRICTED = "#{NEW_PLUGIN_DEST_DIR}/new_3p_plugin_app_restricted"
 
   def test_plugin_setup_and_dirs
     # Make sure the test script has copied the test plugin to the directory
@@ -98,11 +99,41 @@ class JavaScriptPluginLoadingTest < IntegrationTest
       get '/do/new_thirdparty_plugin/test'
       assert_select '#z__ws_content p', 'New contents'
 
+      # ---------- PRIVATE VERSION OF DEPLOYED PLUGIN ----------
+
+      # Create a copy of the plugin which is restricted to this application
+      FileUtils.cp_r("#{NEW_PLUGIN_SOURCE}/.", NEW_PLUGIN_DEST_APP_RESTRICTED)
+      plugin_json = JSON.parse(File.read("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/plugin.json"))
+      plugin_json['restrictToApplicationId'] = _TEST_APP_ID
+      File.open("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/plugin.json","w") { |f| f.write JSON.generate(plugin_json) }
+      File.open("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/template/test.html", "w") { |f| f.write("<p>New contents for app #{_TEST_APP_ID}</p>") }
+      File.open("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/version", "w") { |f| f.write(KRandom.random_api_key) }
+
+      # Load it and check there are two plugins registered
+      do_plugin_reloading
+      assert_equal NEW_PLUGIN_DEST, KPlugin::PLUGINS['new_3p_plugin'].plugin_path
+      assert_equal NEW_PLUGIN_DEST_APP_RESTRICTED, KPlugin::PRIVATE_PLUGINS[_TEST_APP_ID]['new_3p_plugin'].plugin_path
+
+      # Application is now using the private plugin version
+      get '/do/new_thirdparty_plugin/test'
+      assert_select '#z__ws_content p', "New contents for app #{_TEST_APP_ID}"
+
+      # Modify template, but not version, check that cached template is still used
+      File.open("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/template/test.html", "w") { |f| f.write("<p>New contents for app #{_TEST_APP_ID} (2)</p>") }
+      do_plugin_reloading
+      get '/do/new_thirdparty_plugin/test'
+      assert_select '#z__ws_content p', "New contents for app #{_TEST_APP_ID}"
+
+      # Update version in private version, check reloaded
+      File.open("#{NEW_PLUGIN_DEST_APP_RESTRICTED}/version", "w") { |f| f.write(KRandom.random_api_key) }
+      do_plugin_reloading
+      get '/do/new_thirdparty_plugin/test'
+      assert_select '#z__ws_content p', "New contents for app #{_TEST_APP_ID} (2)"
+
     ensure
       # Clean up after the test
-      if File.directory? NEW_PLUGIN_DEST
-        FileUtils.rm_r NEW_PLUGIN_DEST
-      end
+      FileUtils.rm_r NEW_PLUGIN_DEST if File.directory? NEW_PLUGIN_DEST
+      FileUtils.rm_r NEW_PLUGIN_DEST_APP_RESTRICTED if File.directory? NEW_PLUGIN_DEST_APP_RESTRICTED
       KPlugin.uninstall_plugin('thirdparty_plugin')
       KPlugin.uninstall_plugin('new_3p_plugin')
       db_reset_test_data
